Merged PR 653696: Use large buffer for directory enumeration when remoting processes
A bug in ProjFs causes incomplete directory enumeration.

The bug isn’t that the caller’s buffer needs to be larger. The bug is that if there are symlinks in the projected directory contents, some of the stuff we’re doing under the covers may require a larger buffer but we’re not figuring out our own buffer requirement the right way.

Basically, ProjFs services nine different FILE_INFORMATION_CLASS types sent by the user. But internally ProjFs use only four of those types when enumerating the local file system and getting data from the provider, since each of them can hold all the info required by some subset of the nine possible types the caller might send (for example, if the caller queries with FileDirectoryInformation, FileFullDirectoryInformation, FileIdFullDirectoryInformation, or FileIdExtdDirectoryInformation, ProjFs uses FileIdExtdDirectoryInformation internally because that holds a superset of the contents of those other types).

Often, the type ProjFs uses internally is bigger than the one the user sent, so ProjFs needs to allocate a bigger buffer than the one the user sent. After we enumerate the local file system and the provider, ProjFs merges the info together and copy it into the caller’s buffer, omitting any contents the user’s type doesn’t hold. The bug is that if the directory contains symlinks, we may need even more internal buffer than ProjFs calculates. If ProjFs has an insufficiently large internal buffer, the provider will run out of space when it provides a directory entry together with the symlink info. The extra symlink info we get from the provider isn’t returned to the user’s call to NtQueryDirectoryInfo, which is why the caller’s buffer doesn’t really need to be bigger.

This PR also deprecates unused QBuildIntegration flag.

Related work items: #1921558