Merged PR 654143: Use standard .net libraries for extracting zips
When a zip file is created on Windows, GZipLib does not handle nested files properly on Mac (the extraction results in a flat list of all files, with the folder names *in* their names). This apparently started to happen from a given version of GZipLib.

Using the standard C# libraries for the zip case. Keeping GZipLib for the other cases (tar, etc.) which supposedly behave correctly on Mac (and there is no standard support from C# for them)