Merged PR 648592: Add some tolerance for late-joining workers in really early releases
It can happen that the orchestrator decides to early-release a worker that is still not attached, and hasn't shown up to the build yet. When this happens, the Exit call either doesn't get to the worker or it reaches it in a pre-attachment state. In any case it causes the worker to fail with DistributionInactiveOrchestrator, which is ugly. This PR adds a bit of tolerance so the worker can complete the attachment before being disconnected.

Related work items: #1838614