Merged PR 653769: Decrement running pip count in the dispatcher queue in a better place
The running pip count for each queue was being decremented when releasing the dispatcher, but the pip may still be "running" even though we decide to inline it. This makes the number be out of sync with the `RunningOrQueued` pips in the PipQueue, which is decremented only after the `RunnablePip.RunAsync` task completes.

A recent change introduced inlining + dispatcher release of remote MaterializeOutput pips, which decremented the "running" count of the Materialize dispatcher once the local worker finished materialization. But we compute `PipQueue.RunningOrQueued - MaterializeDispatcher.Running` to decide if the scheduler is done, so these numbers being out of sync was making this never happen (as `MaterializeDispatcher.Running == 0` but `PipQueue.RunningOrQueued == {Pending materializations in the workers}`).

This fixes the behavior, and as a side effect corrects the count of the "actually running pips" to include the ones that are inlined (and were releasing the dispatcher).