Merged PR 654144: Stop flushing statsprf.json file upon cancellation
This file gets flushed slightly differently as part of the scheduler's shutdown process. That's good for typical builds, but when a build is cancelled due to timeout that makes it harder to tell what was going on during at the point that it was cancelled. So skip flushing the file if cancellation is requested to preserve its state.