Merged PR 636895: Avoid setting quota request result twice
This PR fixes the following issue:

```
QuotaKeeper: Purge loop failed for 'Reservation request for 1008 bytes' with unexpected error: System.InvalidOperationException: An attempt was made to transition a task to a final state when it had already completed.
   at System.Threading.Tasks.TaskCompletionSource`1.SetResult(TResult result)
   at BuildXL.Cache.ContentStore.Stores.QuotaKeeper.<ProcessReserveRequestsAsync>d__51.MoveNext() in \.\Public\Src\Cache\ContentStore\Library\Stores\QuotaManagement\QuotaKeeper.cs:line 730
```

This may happen when the quota request was successful, but due to a high rate of put operations the system still can be above hard limits.