Merged PR 656572: Avoid doing file existence write checks on detours when possible
Profiler shows a ~10% of the time going into PolicyResult::ShouldForceReadSharing when FileExistenceBased checks are enabled (most of the time goes into GetFileAttributes happening underneath) for selected pips.
This PR makes ShouldForceReadSharing to be only affected by policy-based configurations (which is essentially what's happening today outside JS builds). The change seems to be speeding up Midgard builds by a 5% to 10%.