Merged PR 656870: Use best effort cache to do TryGetReparsePointTarget
Profiling shows that TryGetReparsePointTarget is taking around half of the detours overhead that comes with doing full reparse point resolving. In the past we found that using the cache for this was slowing things down, basically because of lock contention. This PR uses a try_lock strategy for reading into the cache, so lock contention is minimized. This strategy was already being used for writing into the cache.
Besides doing some local experiments that confirms the speed up, a full clean build of Midgard is now taking 10% less with this strategy.