Merged PR 656992: Back to Legacy pipe reader: Pipe read cancellation strikes back
Apparently, the async pipe reader recommended by the .NET team (the one based on the stream reader) also suffers from pipe read cancellation. Because there's no retry mechanism in that pipe reader, and the pipe is still OK, Detours gets stuck in sending messages and so the pip/target becomes non-terminating. This explains the recent spike in non-terminating builds in both BuildXL and QuickBuild.