Merged PR 656995: Add unsafe execute option to disable sandboxing at the pip level
For large javascript prep-builds (the main motivator here being the midgard build), where the build consists mainly of a single (huge) pip that does a lot of operations (e.g. `yarn install`), the overhead of detouring the process (and even of the cache lookup on a cache hit) is large enough that it makes sense to run it outside of the sandbox every time and lose the benefits of caching.

There is already an option to disable Detours for the whole build (that gave good results in the builds in question), but we still want to cache the other pips (i.e., download pips) so this PR adds the option to specify this at the pip level through the unsafe configurations in the spec.