Merged PR 657963: Update Windows to bond 10.0 package
This only updates the Windows packages. The macOS and linux packages need to be created by building from source. We do not perform distributed builds on those platforms so I don't expect a compatibility issue with this change. We are planning to upgrade everything to protobuf anyway.

Related work items: #1938788