Merged PR 657936: [Linux Pipeline] Use more aggressive cache keys
The effects of this change should be:
- a batched CI build is automatically triggered every for new changes pushed to 'master'
- those CI builds produce caches that can be used (restored) in subsequent PR validation builds
- the cache task's `key` and `restoreKeys` properties are updated so that caches are associated with the exact change being built; this promotes restoring the most recently produced cache entry (as opposed to, like it was before, the entry associated with the first build done with the same LKG)

A potential downside of this change is that now we will be caching ~4GB of the `Out/Cache` folder for every single commit we build.  ADO does not enforce any limits on the total cache size ([see](https://docs.microsoft.com/en-us/azure/devops/pipelines/release/caching?view=azure-devops#q-is-there-a-limit-on-the-size-of-a-cache)), so maybe this won't be an issue.  Uploading 4GB of such cached content takes ~2m, so this change should better reduce the BuildXL build time by more than 2 minutes.