Merged PR 657935: Improve PathTree structure for tracking reparse points
When doing full reparse point resolution. profiler shows that a significant amount of time is spent on inserting/retrieving elements into the PathTree, and in particular resolving the less than operation the map at each node requires.
Vectors are known to behave better than maps for lower number of elements. This PR uses a structure for holding the children of a node that starts as a vector and turns into a map when the number of elements reach certain threshold.
Midgard main clean build drops from ~55m to ~47m after this change.