Merged PR 657205: Propagate ServiceVersion, Environment and ConfigurationId to Kusto in out-of-...
This PR fixes the logic used for propagating `Environment`, `ServiceVersion` and `ConfigurationId` properties to kusto in out-of-proc modes.

The first two were just not propagated properly, and for the `ConfigurationId` case the logic was changed to separate the launcher case from the normal out-of-proc case. The launcher still computes the hash of the configuration obtained from a separate git repo and the out-of-proc case uses the same `ConfigurationId` that the parent process uses.

Related work items: #1939096