Merged PR 657217: Mark downstream pips direct dirty/non-materialized even if they are not filtered to in the current graph
Mark downstream pips direct dirty/non-materialized even if they are not filtered to in the current graph.  This is important if a change affects downstream pips which are not currently filtered to, but will be in a future build.
EG: pip B uses outputs from pip A, and pip A is filtered to.  When A runs, it needs to mark B non-materialized so the next build it re-runs.