Merged PR 658746: Fix LimitingResourcePercentages not going to telemetry
This is a fix for the change made in !580983

We have some events that log to discrete events in telemetry but have some fields in their event payload go to the .stats file for sake of logging convenience. ExecutePhaseComplete was one of those events. It would be redundant to log them both to FinalStatistics and the discrete event tables in Kusto, so there's some filtering in the StatisticsEventListener to filter them out when sending FinalStatistics. This is based on a pattern in the stat name. Unfortunately the above change followed the same stat name pattern as when we have a discrete event. The result is that they got logged in the stats file but they never made it to telemetry.