Merged PR 658251: System.NullReferenceException bug during cache init
Added a conditional check to **CacheCoreFingerprintStore.cs** to avoid the above crash, it was added here to ensure there as less code change as possible(avoiding overriding methods at multiple locations). Did not add a condition to **GetCacheEntryAsync** method in **MemoizationStoreAdapterCacheReadOnlySession.cs** class as it may just fix the current issue but may not avoid similar bugs.

Related work items: #1937009