Merged PR 659760: Add an option to resolve machine ids eagerly
This PR adds an option to resolve a list of machine ids eagerly instead of doing the resolution lazily and fail with `InvalidOperationException`.

With the new option, if the resolution fails, the error is logged and the location is skipped.

This PR also has a few performance fixes.

* The new approach generates almost half of the memory traffic. For the performance test added with this PR, the old version generates 9.9Gb and the new one - 5.3Gb (!)
* The PR removes the overhead in `ClusterState` by keeping an internal state as a field and not as a result property.
* The PR also changes the reputation tracking by keeping the map from `MachineId` to the repupation instead of `MachineLocation` to the reputation.
* The new implementation is simpler compared to the old one and I'm going to remove the old one ones the new one is tested and deployed.

Overall, even though the locations are computed eagerly, the new implementation is faster than the old behavior 2-3x(!). See the `PerformanceTest` in this change.

Related work items: #1941877