Merged PR 660014: Use well known symbols for .net core applications
Our sdk was not using standard symbols for the various .net core flavours, for instance, if you create a C# application from Visual Studio you'll get `NETCOREAPP` symbol defined, but not `NET_CORE` one that the bxl sdk defined.

Another thing that we were missing is `NET5_0_OR_GREATER` kind of symbols which are very useful for making more robust changes, because in many cases, when a new feature is added to .net core the upcoming version will continue supporting it.

Related work items: #1944103