Merged PR 659941: Use HashCode.AddBytes for byte[] hash computation
When I was working on [this PR](https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/659760) I've noticed that the hash code computation for `MachineLocation` was quite visible in the profiler. Then I checked that starting from .net6 there is a vectorized version in `System.HashCode` struct for computing a hash code of byte arrays.

I ran a benchmark that computes a hashcode for a byte[] of a different length (10, 100 and 1000 bytes) and here is the results:

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/659941/attachments/image.png)

I also tested it within a dictionary and the new implementation was more efficient as expected.