Merged PR 660642: [AnyBuild][BuildXL-Remoting] Upgrade AnyBuild in BuildXL
Upgrading AnyBuild in BuildXL so that we can use correct symlink resolution when substs are present.

This PR also contains some fix in getting reparse point target. Some bottleneck in AnyBuild is identified due to calling `TryGetReparsePointTarget`. Basically, due to incorrect `FileShare` and `FileDesiredAccess` we get denied access. When that happened, BuildXL tried to find who was holding a handle to the file.

This often happens when we enumerate `C:\` or `C:\Users`.