Merged PR 660367: System. FormatException bug fixes.
The bug was thrown by the String table among the three tables. The problem occurs when the sizeInBytes value is a long value. The sizeInBytes is expected to be long but during deserialization the sizeInBytes is read as ReadInt32Compact() instead of ReadInt64Compact(). Replaced the method with ReadInt64Compact()

Related work items: #1932621