Merged PR 660897: Trace the number of hashes in PlaceBulk
As we discovered today, we don't use `PlaceBulk`, but we do have plans to change that and start using it in the future.

This PR just changes the way this operation is traced by adding the number of files placed by this operation which can be very important for analyzing the effectiveness of the bulk operation.

Related work items: #1946135