Merged PR 662013: Allow configuring servergc heap count and set it to 3 for bxlanalyzer
Allow configuring servergc heap count and set it to 3 for bxlanalyzer.  If not set, the process will use cpu on each logical processor, which is a problem with a lot of processors and a lot of the same process running at once.  3 was found to be the minimum value to give all the benifits for bxlanalyzer on the Word x64 debug project.  Since bxl.exe doesn't run more than one instance at a time, no need to set it