Merged PR 662957: Reduce lock contention in TryUpsert
`RocksDbContentLocationDatabase.TryUpsert` was using the first byte of the serialized strong fingerprint as the key in locks, but the first byte is a length, so the lock distribution was low.

This PR pushes the common functionality into `ContentLocationDatabase` and uses the similar lock index computation in `RocksDbContentLocationDatabase` and `RocksDbMemoizationDatabase`.

Related work items: #1952009