Merged PR 663694: Scrub stale packages by default on ADO builds
This should address the CG issues we get on stale packages from ADO builds.
Requires the nuget resolver changes to become part of the LKG in order to take care of the nuget packages, but that should happen this week.
An equivalent change is coming on the couple of CB queues we use for our selfhost.