Merged PR 662506: Stop using L3 for cache lookup operations when the chances of getting hits are low
Operations against the L3 cache can be expensive. Experiments have shown that on close-to-clean builds, dev cache can making builds considerably slower. This PR adds a `AvoidRemote` hint for the cache operations used in cache lookup: when set to `true`, the vertical aggregator won't use the remote session for that operation.

The hint is set to `true` using a simple heuristic: it is done in cases where the pip is part of a chain of 2 consecutive cache misses. The assumption is that the path is dirty enough that the pip will be a miss in the L3 if is in the L1.

A lot of this PR is just plumbing the `OperationHints` in the appropriate methods, where before just `UrgencyHint` was used. The beef of this PR is in `VerticalCacheAggregatorSession`, `Scheduler`, `PipExecutor`, and the tests in  `RemoteCacheShortCircuitTests` and `VerticalAggregatorTests`

Related work items: #1458668, #1931650