Merged PR 665772: Propagate miss count in a dependency chain even when non-process-pip in the chain are hits
In a chain `PipA <- PipB <- [SealDirectoryPip] <- PipC`, if the seal directory pip is a "cache hit", we are resetting the miss count and so PipC won't get the AvoidRemote hint.

This PR makes it so non-process pips always propagate the miss count, regardless of their result. This means that the non-process pips in the graph will be transparent to the remote cut-off heuristic.

Added a unit test that reproduced the issue.