Merged PR 665403: Add ability to assert existence of files in opaques that are in a different fragment
 Currently this is only used in the context of pip graph fragments so that a fragment can assert a file exists in a directory which originates in another fragment.

Each of these file artifacts, upon deserialization from the fragment, need to be added to the graph after the directory they come from is added and before the file is used.

Since the assertions are loaded before any pips from the fragment, only directories in other fragments can be specified.

Future work can load the assertions in topological order with the pips that reference them so they can reference a directory created in the same fragment.