Merged PR 665954: Add merge operators to ContentLocationDatabase
This PR adds merge operator support to `RocksDbContentLocationDatabase`.

The design is based on merge operator support in RocksDb and adds "mergable" machine id set called `LocationChangeMachineIdSet` that allows keeping not only additions but also removals of the locations.

## Performance differences
The performance is roughly 40% better based on the manual performance test added in this PR.

Checking the performance and correctness in CBTest as well.