Merged PR 664906: Fix underbuild due to dirty check.
Remove check for already dirty/materialized.  Technically the first one is correct, but the second one was missing the materialized check.  Given that the consequences of getting this wrong are high (underbuilds) and the cost to dirtying some pips multiple times are low, let's remove the check entirely.

This happens when a graph change causes a pip to be direct dirty, causing  downstream pips to be dirty.  Then, if the downstream pip input is also changed, that downstream pip would not be marked direct dirty because it was already dirty.