Merged PR 664242: Categorized Proactive Copy Telemetry
Categorized Proactive Copy Telemetry.

We currently get a lot of "Ring=[Unknown]" status in ProactiveCopyIfNeededAsync telemetry. This is very unhelpful in tracking down issues, and appears to often occur when no copies are event attempted. we were creating new PushFileResults with a string constructor and no specified status, which then defaults to Status=Unknown.

This PR:
1. Adds new Proactive copy status to understand copy failure reasons.
2. If all machines have the copy we mark it as success but log additional string that copy was skipped
3. "Could not find other build machines" and "Could not find build" is marked as rejected but logged the reason why it was rejected

The log string to remain uniform for all possible outcome:

**Success Count=[0/1/2] InsideRing=[Status/Not Applicable], OutsideRing=[Status/Not Applicable] InsideRingError=[error] OutsideRingError=[error]**

Related work items: #1947887