Merged PR 667063: Fail pips faster when a connection is lost with a worker
This is the scenario we are fixing:
We lost the conneciton with one of the workers during metabuild. however, there are already 50k pip build requests added to m_buildRequests for this worker. those are for materializing outputs. As we send 1k pips at max in one grpc message, we have to spend 20 minutes (10minute deadline per call and 2 attempts) for 1k pips. so, the metabuild times out