Merged PR 669257: Don't use the NuGet cache when downloading the BuildXL LKG on local builds
NuGet uses a shared non-deduplicated cache by default. On local machines this results in a lot of space taken by stale bxl LKGs. Skip the cache for this case to save on disk space. The only scenario that a shared cache could be helping with in this case is if a developer has multiple bxl enlistments, but this case is not super common, and it will only save download time (not space).