Merged PR 669232: Handle index out of bounds in Windows Error Reporting code
It seems the stack trace cannot always be retrieved. Only include it in the WER report if it is available

```
System.AggregateException: Failed to create custom Windows Error Reporting report (Index was outside the bounds of the array.) (A Task's exception(s) were not observed either by Waiting on the Task or accessing its Exception property. As a result, the unobserved exception was rethrown by the finalizer thread. (Unable to read data from the transport connection: An existing connection was forcibly closed by the remote host..))
 ---> System.IndexOutOfRangeException: Index was outside the bounds of the array.
   at BuildXL.WindowsErrorReporting.ComputeParameters(Exception ex, BuildInfo buildInfo, String sessionId) in \.\Public\Src\App\Bxl\WindowsErrorReporting.cs:line 198
   at BuildXL.WindowsErrorReporting.CreateDump(Exception ex, BuildInfo buildInfo, IEnumerable`1 filesToAttach, String sessionId) in \.\Public\Src\App\Bxl\WindowsErrorReporting.cs:line 42
```