Merged PR 669803: [VM] Introduce temp folder shared by pips executed in VM
When pips are executed in VM, some tools cannot be executed with `net use`, and thus those tools need to be copied to a temporary location in the VM. The pip can use its specified temp folder as this temporary location. But since pips's temp folder is usually unique, it means that the same tool will be copied multiple times to the VM if there are multiple pips needing that tool. Not only that it will degrades performance, the VM will run out of space quickly because it has many pips running there at the same time, and all of them need a large tool.

With this shared temp folder, one pip can copy the tool once, and other pips can simply reuse it.