Merged PR 670363: Changes for code base property for BXL telemetry
Added codebase property to gather information about the branch triggering the build.
In CB this information is passed through traceinfo.
In ADO this information is extracted from the predefined variable "BUILD_REPOSITORY_NAME".
Added two unit test cases, one to test if the changes are effective in the ADO environment.
The other is to test the changes for a CB environment and also to check if the user passed value overrides the ADO set value or not.

Related work items: #1954489