Merged PR 670962: Copy cancellation based on Network Copy Speed
Currently copy speed is calculated based on the number of bytes copied in a certain interval. This interval includes both network and disk copy. We want to cancel a copy if network is slow, not if disk is busy. To achieve that we need to compute network copy speed.

Related work items: #1959303