Merged PR 671092: Changes for build entity property for BXL telemetry.
Added buildEntity property to gather information about the build queue(CB env) and build pipeline(ADO env).
In CB this information is obtained from "cloudBuildQueue"  name passed via traceInfo
In ADO this information is obtained from the ADO predefined variable "SYSTEM_DEFINITIONID".
Added two unit test cases, one to test if the changes are effective in ADO environment.
The other is to test the changes for a CB environment and also to check if the user passed value overrides the ADO set value or not.
Used the name "buildEntity" to represent both build queue and build pipeline. But suggestions are welcomed.
Another thought here is that. Since we already have **cloudBuildQueue** as a part of the env string in the telemetry. Would it be enough to just add another property called **pipeline** to represent this information for ADO. That way we can avoid duplication of information if it is CloudBuild.

Related work items: #1948523