Merged PR 672371: Fix use of shared VM temp folder
The content of shared VM temp folder gets cleaned before the pip run. That should not happen because the temp folder is shared by multiple pips. This PR only ensures that the shared VM temp folder exists without cleaning it up when it's already existed.

This PR also enables the unit test that cannot be enabled previously due to waiting for a new LKG.