Merged PR 672591: Fix machineId merge
Fix machineId merge.

This PR also moves away from regular lock (based on `Monitor.Enter`) to use RW lock.
When the merge operators are used we can just acquire read lock when writing a merge, and only acquire a write lock on Delete. But when the merge operators are off, then a write lock is used where we used to have a normal lock.

The change should not affect the performance, because the locks with no contention are pretty comparable in terms of perf.

Here are the results of a benchmark:
```
|           Method |     Mean |    Error |    StdDev |
|----------------- |---------:|---------:|----------:|
|  NodeWithFatLock | 89.30 ms | 4.296 ms | 12.599 ms |
|   NodeWithNoLock | 17.07 ms | 0.333 ms |  0.945 ms |
| NodeWithThinLock | 17.58 ms | 0.595 ms |  1.754 ms |
|   NodeWithRwLock | 32.35 ms | 0.495 ms |  0.661 ms |
```

Fat lock is a lock when `GetHashCode` was called on an instance that was passed to `Monitor.Enter`. Note, then RW lock is slower but comparable to a normal (thin) lock.

The benchmark was taking about 5K locks.