Merged PR 670540: Conceptual redesign of the IpcMoniker abstraction
Conceptual redesign of the IpcMoniker abstraction. There are no behavioral changes, this is only a change in the design that is also adjusted to the current usage of the previous design. "IPC monikers" are not a black box created by an `IIpcProvider` anymore, but only generic handles (string wrappers) to a connection string: thus, we now use the simple `IpcMoniker` struct for them instead of `IIpcMoniker`s. and the `IIpcProvider` contract changes as to only require that a connection string will be stable for a given moniker (usage of a connection string was already guaranteed to happen after rendering it through `RenderConnectionString`, so any special behavior / memoization can happen at that point instead of in `CreateMoniker` or `LoadMoniker`... in fact this was already the case).

This is motivated by recent changes in the layout of the related assemblies: the actual usage of monikers was exactly this (via the `StringMoniker` implementation). Here's a comment from Aleks in my last PR with more context.

_The original intention was that you get an IpcProvider, you call its CreateMoniker() method and it gives you back something opaque, encoding whatever information it may want in the created moniker. Later, you can return the same moniker to the same provider and the provider will be able to decode whatever information it put into it:_.

_The original use case for that was to encode a connection string (which, at its simplest form, is just a port number). The problem with that approach was that eagerly deciding on a concrete connection string (e.g., a port number) can later lead to unnecessary collisions (e.g., the port might not be available any longer). So we moved to monikers being just GUIDs and providers having to maintain a separate map from a moniker (guid) to a concrete connection string; that way a connection string can be tied to a moniker as lazily as possible._

_And it looks like that is the correct design for this: a moniker is nothing but a guid (i.e., it may not encode any concrete information). With that design, no IIpcMoniker abstraction is needed._ _

The relevant changes are in `IIpcProvider` and `IpcMoniker`, the rest of the changes are just renaming.

Related work items: #1969773