Merged PR 672992: Add performance-optimized extension methods for 'IReadOnlyList'
This PR adds a set of extension methods for `IReadOnlyList<T>` to avoid using the extension methods from `Enumerable.XXX`.

The PR also fixed 'CA1826' rule that warns when the `Enumerable.Count/Last/First` is used on `IReadOnlyList<T>`.

The PR also switches away from `RequiresNotNull` to `Requires(x != null)` because the latter is a prefered way for checking for nulls.