Merged PR 673542: Change Ninja resolver settings to intervene CXX command lines into an environment setting
Controlling this behavior, which is basically a hack, doesn't belong in the resolver settings. We want to come up with a design to avoid doing these things. This PR lets us toggle the logic via an environment setting for experimenting.

This PR includes no behavioral changes (in the default case where the env variable is not set).