Merged PR 673619: Add stack trace for debugging exception in saving incremental scheduling state
PendingNodes of dirty node tracker should no longer be updated when it is time to save the incremental scheduling state. However, bug #1948263 indicates that the PendingNodes is still being updated.

In this PR we throw an exception when an update to PendingNodes happens after they are applied to the dirty node tracker.

The exception is caught during saving of incremental scheduling state and is logged to telemetry table `IncrementalSchedulingSaveState`

Related work items: #1948263