Merged PR 674113: Fix unique/total bytes logging output if the root is a single chunk
I was playing with 1MB average chunk value and found out a small bug where if the file is < 1MB, the chunker creates a single chunk as expected, but the logging for unique and total bytes is displayed as 0.

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/674113/attachments/image.png)

Added an else condition to count the transitive content bytes for this scenario.

After:
![image (2).png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/674113/attachments/image%20%282%29.png)