Merged PR 674879: Exclude non essential methods from exception's stack trace
This PR updates `Ben.Demystifier` logic that is embeded into `BuildXL.Utilities` to support the latest language changes.

The PR also adds support for respecting `StackTraceHiddenAttribute` for async methods and adds tests to check that the methods marked with `StackTraceHiddenAttribute` are not part of the error output.

Related work items: #1979115