Merged PR 677789: Clear the pool of PageStreams after the serialization is done
During the graph construction, we try to reuse the memorystreams which do not use more than 90% of their capacity. However, when we are done with serializing the pips, we forget to clear the pool of those memorystreams. We keep them in memory during the execute phase till the end of the build.

In one of the office dev builds, this was taking 4gb: 30% of the total memory usage.

![Screenshot 2022-09-01 121314.png](https://mseng.visualstudio.com/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/677789/attachments/Screenshot%202022-09-01%20121314.png)