Merged PR 678113: [BuildXL Perf Bash] Reduce boxing allocations by implementing IEquatable interfaces
This PR reduces about 1Gb of allocations per minute for COSINE builds and its a bit more interesting than the others.

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/678113/attachments/image.png)

A lot of allocations come from 'GetHashCode' and 'Equals' methods because some types does not implement `Equals` and `GetHashCode` causing the runtime to box the structs on these calls.

If you want to read more on this topic, please have a look at [The performance implications of default struct equality in C#](https://devblogs.microsoft.com/premier-developer/performance-implications-of-default-struct-equality-in-c/)