Merged PR 678711: Ignore Lage bogus dependencies and add tests
For the case where project A depends on B, A has a 'build' verb but B doesn't, Lage generates a graph where A#build depends on B#build, but B#build doesn't actually exist as a node. Let's ignore that dependency instead of bailing out.

In addition, add some scheduling tests for Lage to validate an e2e run of the resolver (and also validate this newly added behavior)

Related work items: #1980091