Merged PR 678656: Set MemoryStream length before returning it back to pool
This PR fixes issue that happened in a few other PRs where I started using pooled `MemoryStream` instances.

The issue with the pool was that it was setting the position but not the length. Meaning that the instances from the pool had old random data that were observable by the new users.

I benchmarked and the new logic is as fast as the old one, i.e. `SetLength` is similar in terms of memory and perf to just setting the position.