Merged PR 679153: Use channels in LoggingQueue for performance reasons
This PR moves `LogginqQueue` to BuildXL.Utilities, adds tests to it and changes implementation from using `BlockingCollection` to use `System.Threading.Channels` for performance reasons.

Here is the micro benchmark that shows 2x perf differences when adding 1, 10 and 1000 to `BlockingCollection` and to `Channel`:

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/679153/attachments/image.png)

And here is a profiling snapshot that shows that such optimization matters.

![image (2).png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/679153/attachments/image%20%282%29.png)

(even though this is not the top method, its high enough. Plus most of the other methods in the list are blocking threads and waiting, that's why they spent so much time in them.