Merged PR 671560: Handle incremental tools that call NtQueryDirectoryFile
Currently a blind spot that leads to underbuilds in droid builds.  When java enumerates a directory with NTQuerydirectory, BXL doesn't add the enumerated files to the list of reads.  It only does so for findfirstfile/findnextfile.

This PR adds EnumerationProbe accesses for files inside the enumerated directory. Thus, they will get the same treatment as findnextfile.

TODO:
- [X] Unit tests :-)