Merged PR 679866: Script to validate SBOM generated in Cloudbuild after build completes
Added a script file to validate SBOM in PR or Rolling pipelines.
The script checks if all the sections in SBOM are present as expected and also throws an error when any of the packages are missing.
The changes have been tested in a PR validation(Primary validation pipeline)
https://dev.azure.com/mseng/Domino/_build/results?buildId=18325311&view=logs&j=a4f1cc8a-0ccb-5e90-36f4-f06f41d0ce59&t=9a294ced-119e-5ffa-e3f7-0c39789dd8e4

Related work items: #1967422