Merged PR 680922: Terminate builds on first internal error in CloudBuild
Continuing build on internal or infrastructure errors in CloudBuild is not helpful.

They waste resources as they progress slowly despite being destined to fail anyway. This can make it take much longer for the user to see the final result or get to the internal error retry build. We see a lot of timeout builds with an upstream internal error. Moreover, they complicate getting an accurate internal error signature for the build if they get externally killed by cmdagent.

The downside to this strategy is that builds with no internal error retry configured could have been run further to completion, potentially exposing more actual user errors. I don't see this as outweighing the problems with continuing to execute these builds.

Related work items: #1988195