Merged PR 680832: [BuildXL Perf] Reduce the size of EnvironmentVariable struct from 48 to 32
An array of `EnvironmentVariable` structs is created per pip and such arrays occupy quite a bit of memory. Here is an example from bxl dump for COSINE:

```
305412   2828539152 BuildXL.Pips.Operations.EnvironmentVariable[]
```

I.e. the arrays are almost 3Gb.

This PR reduces the size of `EnvironmentVariable` struct from 48 to 32 bytes by flattening types used in it.

It also adds a reference to `ObjectLayoutInspector` nuget package that allows inspecting the layout at runtime.

The layout before:
![image (2).png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/680832/attachments/image%20%282%29.png)

Here is the layout after:
![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/680832/attachments/image.png)