Merged PR 681483: Add linux support for capturing process dumps
Add linux support for capturing process dumps. First version, this will likely be improved in subsequent iterations to make it more robust.
* The overall flow assumes dumps going to /var/crashes, which is the most typical location. There is a TODO for generalizing this
* There are some special tweaks targeted to apport, the default process handler in Ubuntu. Things should work without apport as well.

Capture the whole process tree is not 100% bulletproof on linux: sending SIGABRT may or may not reach child processes (depends on how the process was spawned). We start with the leaves as a way to improve the chances to capture more meaningful stuff.
Tests could be added in the future, but let's give this feature some mileage first/